requirements.txt
pandas
matplotlib
astropy
torch
scipy
python-dotenv
requests
fastapi
uvicorn
scikit-learn
